+++ packages/types/src/index.ts
@@ -1,8 +1,16 @@
-export type {
+import type {
   Char,
   CharSet,
 } from "./characters";
-export type {
+import type {
+  AttributeOptions,
+  CssDeclarationPropertyOptions,
+  CssDeclarationValueOptions,
+  MultiValueAttributeOptions,
+  QuerySelectorOptions,
+  SingleValueAttributeOptions,
+} from "./options";
+import type {
   MangleEngineOptions,
   MangleExpression,
   MangleExpressionOptions,
@@ -13,3 +21,23 @@ export type {
   WebManglerOptions,
   WebManglerPlugin,
 } from "./types";
+
+export type {
+  AttributeOptions,
+  Char,
+  CharSet,
+  CssDeclarationPropertyOptions,
+  CssDeclarationValueOptions,
+  MangleEngineOptions,
+  MangleExpression,
+  MangleExpressionOptions,
+  MangleOptions,
+  MultiValueAttributeOptions,
+  QuerySelectorOptions,
+  SingleValueAttributeOptions,
+  WebManglerEmbed,
+  WebManglerFile,
+  WebManglerLanguagePlugin,
+  WebManglerOptions,
+  WebManglerPlugin,
+};